Show favorites on profile page if any have been made.

+++ inkscape/resource/views.py
@@ -275,10 +275,12 @@ def view_resource(request, item_id):
 @login_required
 def like_resource(request, item_id, like_id):
     item = get_object_or_404(Resource, id=item_id)
+    if item.user.pk == request.user.pk:
+        raise Http404
     (like, is_new) = item.votes.get_or_create(voter=request.user)
     if not is_new and like_id == '-':
         like.delete()
-    elif like_id == '+':
+    elif like_id == '+' and item.user.pk != request.user.pk:
         like.vote = True
         like.save()
     return redirect("resource", item_id)
